Move salt.utils.mkstemp() into salt.utils.files
Also use a default prefix
