+++ salt/utils/files.py
@@ -9,6 +9,7 @@ import logging
 import os
 import shutil
 import subprocess
+import tempfile
 import time
 
 # Import salt libs
@@ -20,6 +21,26 @@ from salt.exceptions import CommandExecutionError, FileLockError, MinionError
 from salt.ext import six
 
 log = logging.getLogger(__name__)
+
+TEMPFILE_PREFIX = '__salt.tmp.'
+
+
+def mkstemp(*args, **kwargs):
+    '''
+    Helper function which does exactly what `tempfile.mkstemp()` does but
+    accepts another argument, `close_fd`, which, by default, is true and closes
+    the fd before returning the file path. Something commonly done throughout
+    Salt's code.
+    '''
+    if 'prefix' not in kwargs:
+        kwargs['prefix'] = TEMPFILE_PREFIX
+    close_fd = kwargs.pop('close_fd', True)
+    fd_, fpath = tempfile.mkstemp(*args, **kwargs)
+    if close_fd is False:
+        return (fd_, fpath)
+    os.close(fd_)
+    del fd_
+    return fpath
 
 
 def recursive_copy(source, dest):
